[hashlib.cc] - Including node.h instead of ev.h and eio.h. node.h includes these files correctly, which might move around from one version of node to the other. (So basically, node.h should be included instead of specifically including ev and eio.) [README.md, makefile, package.json, test.js] - Changing the '/build/default/hashlib' path to '/build/Release/hashlib'. Node-waf, starting from node 0.5.5 uses the Release environment by default.
