+++ makefile
@@ -5,7 +5,7 @@ tests:
 	nodejs ./test.js
 
 install:
-	@mkdir -p ~/.node_libraries && cp ./build/default/hashlib.node ~/.node_libraries/hashlib.node
+	@mkdir -p ~/.node_libraries && cp ./build/Release/hashlib.node ~/.node_libraries/hashlib.node
 
 all: build install
